% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ath_rankings.R
\name{ath_ranking}
\alias{ath_ranking}
\title{Generate Rankings}
\usage{
ath_ranking(.ratings, .season = max(.ratings$season),
  .date = as.character(Sys.Date()), .prov = 6)
}
\arguments{
\item{.ratings}{Raw ratings for all skiers}

\item{.season}{season}

\item{.date}{date}

\item{.prov}{provisional number of races}
}
\description{
Generate Rankings
}

