% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr_elo_funs.R
\name{calc_elo_spr}
\alias{calc_elo_spr}
\title{Calculate Sprint ELO Ratings}
\usage{
calc_elo_spr(races, current_rating, K = 32, P = 1.25, provisional_n = 6L,
  season_shrink = 2/3, default_rating = 1300)
}
\arguments{
\item{races}{list of data.frame's with race results, in chronological order}

\item{current_rating}{data.frame with pre-initialized starting rating}

\item{K}{K factor}

\item{P}{inflation factor for provisional ratings}

\item{provisional_n}{number of races a skier is considered provisional}

\item{season_shrink}{amount to shrink ratings between seasons}
}
\description{
Calculates entire history of ELO ratings for sprint races.
}

