% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/dst_elo_funs.R
\name{calc_elo_dst}
\alias{calc_elo_dst}
\title{Calculate Distace ELO Ratings}
\usage{
calc_elo_dst(races, current_rating, K, P, provisional_n, season_shrink,
  default_rating = 1300)
}
\arguments{
\item{races}{list of data.frames with results, in chronological order}

\item{current_rating}{data.frame with pre-initialized starting rating}

\item{K}{List of K factors for each race type}

\item{P}{inflation factor for provisional ratings}

\item{provisional_n}{number of races a skier is considered provisional}

\item{season_shrink}{amount to shrink ratings between seasons}

\item{default_rating}{}
}
\description{
Calculates entire history of ELO ratings for distance races.
}

